Moves Tokens class into Token header
