+++ src/lexer/Token.hpp
@@ -8,9 +8,12 @@
  */
 
 #pragma once
+#include "Exceptions.hpp"
+#include <queue>
 #include <string>
 
-/* Stores the tokens read from the user's file */
+
+/* Stores a token */
 class Token : public std::string
 {
 public:
@@ -47,3 +50,40 @@ private:
     /* Type of the token */
     Type _type{Type::NotSet};
 };
+
+
+/* Stores all tokens */
+class Tokens : private std::queue<Token>
+{
+public:
+    /* Pops and returns the token at the front of the queue */
+    [[nodiscard]] inline Token dequeue();
+
+    /* Returns the front token */
+    using std::queue<Token>::front;
+
+    /* Skip token */
+    using std::queue<Token>::pop;
+
+    /* empty() */
+    using std::queue<Token>::empty;
+
+protected:
+    friend class Tokenizer;
+
+    using std::queue<Token>::push;
+};
+
+
+Token Tokens::dequeue()
+{
+    if (empty())
+    {
+        ThrowException("Token queue is empty!");
+    }
+
+    Token next = front();
+    pop();
+
+    return next;
+}
